Password_generator/requirements.txt
Pillow
customtkinter
